This is useful for measuring the CPU performance impact of changes on
the BE. See the file for examples of measuring the overhead of QP
calls.

As an illustration, I measured QP performance before and after tracking
`field-usage`. The output looks like this:

```
Running the QP (with *execute* stubbed out) over 4 rounds of 10 calls each

|  Baseline | Baseline (each) | Field Usage | Field Usage (each) | Ratio |
|-----------+-----------------+-------------+--------------------+-------|
| 563.188ms |        56.319ms |  1554.700ms |          155.470ms | 2.761 |
| 540.039ms |        54.004ms |  1504.453ms |          150.445ms | 2.786 |
| 546.126ms |        54.613ms |  1492.443ms |          149.244ms | 2.733 |
| 528.079ms |        52.808ms |  1503.259ms |          150.326ms | 2.847 |
:done
```